src/requirements.txt
Django==3.0.4
djangorestframework==3.11.0
Faker==4.0.2
murmurhash3==2.3.5